refactor STFPM implementation
